Refactoring: `lib/util/editNotes.ts` -> `lib/MB/EditNote.ts`
